Bug Fix - frequency change unreachable
BUG FIX: Managed to put the call to change frequency inside the section
of conditional that only runs when the encoder pushbutton is pressed.
Since this is an extremely infrequent UI event, the signal generator
would seem completely unresponsive, until the button is pressed, at
which point an ungodly large number of accumulated frequency change
counts will execute.